Add BaseWidget with get_context and get_form
