+++ src/senaite/core/z3cform/widgets/basewidget.py
@@ -1,0 +1,75 @@
+# -*- coding: utf-8 -*-
+#
+# This file is part of SENAITE.CORE.
+#
+# SENAITE.CORE is free software: you can redistribute it and/or modify it under
+# the terms of the GNU General Public License as published by the Free Software
+# Foundation, version 2.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+# details.
+#
+# You should have received a copy of the GNU General Public License along with
+# this program; if not, write to the Free Software Foundation, Inc., 51
+# Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+#
+# Copyright 2018-2023 by it's authors.
+# Some rights reserved, see README and LICENSE.
+
+from bika.lims import api
+from plone.z3cform.fieldsets.interfaces import IDescriptiveGroup
+from z3c.form.interfaces import ISubForm
+from z3c.form.widget import Widget
+from zope.component import getUtility
+from zope.component.interfaces import IFactory
+
+
+class BaseWidget(Widget):
+
+    def get_form(self):
+        """Return the current form of the widget
+        """
+        form = self.form
+        # form is a fieldset group
+        if IDescriptiveGroup.providedBy(form):
+            form = form.parentForm
+        # form is a subform (e.g. DataGridFieldObjectSubForm)
+        if ISubForm.providedBy(form):
+            form = form.parentForm
+        return form
+
+    def get_context(self):
+        """Get the current context
+
+        NOTE: If we are in the ++add++ form, `self.context` is the container!
+              Therefore, we create one here to have access to the methods.
+        """
+        schema_iface = self.field.interface
+        if schema_iface and schema_iface.providedBy(self.context):
+            return self.context
+
+        # we might be in a subform, so try first to retrieve the object from
+        # the base form itself first
+        form = self.get_form()
+        portal_type = getattr(form, "portal_type", None)
+        context = getattr(form, "context", None)
+        if api.is_object(context):
+            if api.get_portal_type(context) == portal_type:
+                return context
+
+        # Hack alert!
+        # we are in ++add++ form and have no context!
+        # Create a temporary object to be able to access class methods
+        if not portal_type:
+            portal_type = api.get_portal_type(self.context)
+        portal_types = api.get_tool("portal_types")
+        fti = portal_types[portal_type]
+        factory = getUtility(IFactory, fti.factory)
+        context = factory("temporary")
+        # mark the context as temporary
+        context._temporary_ = True
+        # hook into acquisition chain
+        context = context.__of__(self.context)
+        return context
